+++ utility/src/main/java/com/tingshulien/game/utility/scorer/Scorer.java
@@ -4,6 +4,6 @@ import com.tingshulien.game.utility.AIContext;
 
 public interface Scorer<T extends AIContext> {
 
-    float score(T context);
+    double score(T context);
 
 }
